(slightly) Improved the text generation algorithm
I changed some things so that the text generation would start and end
quotes properly, instead of randomly inserting the quotations
everywhere. This works for the most part, but results in crashes
sometimes, or results in more nonsense sentences. It's good enough for
now but I'll need to go back in the future.